Add eager vars and move builtin ones out of environ.lua
